remove app tests (#3160)
