+++ .github/workflows/ci.yml
@@ -100,27 +100,8 @@ jobs:
           go-version: ${{ env.go-version }}
       - name: setup env
         run: make
-      - name: unit tests (except app test)
-        run: make test-no-app-test
-        env: 
-          TEST_LOG_LEVEL: debug
-  apptests:
-    runs-on: ubuntu-latest
-    needs: filter-changes
-    if: ${{ needs.filter-changes.outputs.nondocchanges == 'true' }}
-    # should take around 7-9 mins
-    timeout-minutes: 20
-    steps:
-      - name: checkout
-        uses: actions/checkout@v2
-      - name: set up go
-        uses: actions/setup-go@v2
-        with:
-          go-version: ${{ env.go-version }}
-      - name: setup env
-        run: make
-      - name: app test
-        run: make test-only-app-test
+      - name: unit tests
+        run: make test
         env: 
           TEST_LOG_LEVEL: debug
   # checkpoint and print a single, clean status message to slack
@@ -132,11 +113,10 @@ jobs:
       - quicktests
       - golangci-lint
       - unittests
-      - apptests
     runs-on: ubuntu-latest
     env:
       # short-circuit success if no non-doc files were modified
-      status: ${{ (needs.filter-changes.outputs.nondocchanges == 'false' || (needs.quicktests.result == 'success' && needs.golangci-lint.result == 'success' && needs.unittests.result == 'success' && needs.apptests.result == 'success')) && 'success' || 'failure' }}
+      status: ${{ (needs.filter-changes.outputs.nondocchanges == 'false' || (needs.quicktests.result == 'success' && needs.golangci-lint.result == 'success' && needs.unittests.result == 'success')) && 'success' || 'failure' }}
     # this is the easiest way to access success/failure state of previous jobs in this workflow
     steps:
       - uses: act10ns/slack@v1
